requirements.txt
logging==0.4.9.6
piexif==1.1.3
pillow==11.0.0
pytz==2024.2
